% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log1ps.R
\name{expm1s10_ceil}
\alias{expm1s10_ceil}
\title{Signed \code{expm1s} With Base 10 rounded to the next upper integer}
\usage{
expm1s10_ceil(x)
}
\arguments{
\item{x}{Numerical vector.}
}
\value{
Numeric vector.
}
\description{
Signed \code{expm1s} with base 10 rounded to the next upper integer.
}
\details{
Compute \code{ceiling(expm1s10(x))}. This is usually used when converting
\code{log1ps10} for labels of axis in a plot. Otherwise we get, for example,
999 instead of 1000 on the axis.
}
\examples{
x <- c(-2, -1, -1 / 10, -0.1 / 10, 0, 0.1 / 10, 1 / 10, 1, 2)
y <- expm1s10_ceil(x)
}
\seealso{
\link{expm1s10}
}
