% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flrpkgTools-package.R
\docType{package}
\name{flrpkgTools}
\alias{flrpkgTools-package}
\alias{flrpkgTools}
\title{Tools Used by Ephel}
\description{
Tools used by Ephel
}
\section{flrpkgTools functions}{

The functions ...
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FrankLef/flrpkgTools}
  \item Report bugs at \url{https://github.com/FrankLef/flrpkgTools/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
