% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print message that begins with "Hello" to console.}
\usage{
hello(msg = "Hello, world!")
}
\arguments{
\item{msg}{Message to print.}
}
\value{
Message, invisibly.
}
\description{
Print message that begins with "Hello" to console.
}
\examples{
hello()
}
