% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_access.R
\name{load_access}
\alias{load_access}
\title{Load Data From MS Access}
\usage{
load_access(dir = getwd(), db = NULL, qries = NULL, env = .GlobalEnv)
}
\arguments{
\item{dir}{directory name of the MS Access db}

\item{db}{file name of the MS Access db}

\item{qries}{query names in the form "df_name" = "qry_name"
where df_name is the name that will be assigned to the data.frame
and qry_name is the query name in the db.}

\item{env}{The environment where the data will be located. Default is
\code{.GlobalEnv}. Don't change that unless you know why you are doing it.}
}
\value{
nb of queries processed
}
\description{
Load data from an MS Access database.
}
\details{
Extract the data from a local MS Access database using a list of queries.
}
