% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalz.R
\name{normaliz}
\alias{normaliz}
\title{Normalize Data Depending on Some Criteria.}
\usage{
normaliz(
  data,
  cols,
  id_cols,
  scale = 1,
  base,
  inverse = FALSE,
  keep = FALSE,
  base_amt = "base_amt"
)
}
\arguments{
\item{data}{Data.frame to normalize.}

\item{cols}{Names of the columns to normalize.}

\item{id_cols}{Names of columns used to join \code{data} and \code{base}.}

\item{scale}{Number used to multiply the resulting data.}

\item{base}{Data.frame containing the base amounts used to normalize.}

\item{inverse}{FALSE (default): Normalize the data. TRUE: Inverse the
normalization.}

\item{keep}{FALSE (default): Don't keep the column with the base amount.
TRUE: keep it.}

\item{base_amt}{Name of the column where the base amount is. Defaults is
"base_amt".  This name must NOT be in the column names of \code{data}.}
}
\value{
Data.frame with normalized data.
}
\description{
Normalize data depending on some criteria.
}
\details{
The data will be normalize using a dataframe, the \code{base} that will be
joined. Typically, for example, the \code{base} would be a summary of sales
by year and entity. Each column is normalized using the formula
\code{data$cols * scale / base}.
}
\examples{
\dontrun{
TODO
}
}
