% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print "Hello" message to console}
\usage{
hello(msg = "Hello, world!")
}
\arguments{
\item{msg}{Message to print.}
}
\value{
Message, invisibly.
}
\description{
Print message that begins with "Hello" to console.
}
\details{
This function is used for testing. The setup of \code{rnlang::abort()}
inside is useful.
}
\examples{
hello()
}
